% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_twokenizer.R
\name{download_twokenizer}
\alias{download_twokenizer}
\title{Download tagger}
\usage{
download_twokenizer(outputLoc, type = c("base", "penn", "irc"))
}
\arguments{
\item{outputLoc}{Location to put the tagger.}

\item{type}{What model to download. "base" must be downloaded first before the other two (i.e. "penn" and "irc") can be downloaded. The Penn model uses Penn Treebank-style tags, which may facilitate working with other packages, whereas the irc model is, as the ArkTweetNLP website states, "a model trained on the NPSChat IRC corpus, with a PTB-style tagset".}
}
\description{
Download Ark-TweetNLP's POS tagger, and optionally, its other models.
}
\examples{
#download_twokenizer("~/Documents", "base")

#MUST DOWNLOAD "base" FIRST:
#download_twokenizer("~/Documents", "penn")
#download_twokenizer("~/Documents", "irc")
}
