% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twokenizer.R
\name{twokenizer}
\alias{twokenizer}
\title{POS tag tweets}
\usage{
twokenizer(tagger_location, text, model_location = NULL)
}
\arguments{
\item{tagger_location}{The location of the "ark-tweet-nlp-0.3.2" folder.}

\item{text}{Tweet to be tagged.}

\item{model_location}{The location of an optional model to be used (e.g. Penn Treebank-style POS tags), instead of the standard model.}
}
\description{
POS tag tweets in R, returning a dataframe with the columns of word, tag, and confidence.
}
\examples{
#twokenizer("~/ark-tweet-nlp-0.3.2",
#"ikr smh he asked fir yo last name so he can add u on fb lololol")
}
